use Int64 to avoid casting

+++ test/Bench.jl
@@ -57,7 +57,7 @@ time_cl(k, data) = @time map(x -> ClausenFunctions.cl(k, x), data)
 n = 100_000
 data = (x_max - x_min)*rand(Float64, n) + x_min*ones(n)
 
-for k in UInt64(2):UInt64(6)
+for k in 2:6
     println("Benchmarking cl($(k),x)::Float64")
     map(x -> ClausenFunctions.cl(k, x), data) # trigger compilation
     time_cl(k, data)            # trigger compilation
